refactor(ricer::locate)!: move locator logic out of `manager` module
Move private sub-module `ricer::manager::locator` out as a public top-level
module named `ricer::locate`. This refactor is apart of a larger effort to
further simplify the `ricer::manager` module in order to prevent it from
becoming a god module.
